requirements.txt
mypy
pytube3